+++ NEWS.md
@@ -1,3 +1,11 @@
+# admiral 0.12.0
+
+## New Features
+
+## Updates of Existing Functions
+
+- The functions `derive_param_bmi()` and `derive_param_bsa() are updated to have the option of producing more values at visits when only weight is collected (#1228).
+
 # admiral 0.11.0
 
 ## New Features